requirements.txt
Flask>=3.0.0
# TensorFlow 2.21+ needs protobuf>=6.31.1; pin so `pip install google-generativeai` does not pull protobuf 5.x and break TF.
protobuf>=6.31.1,<8.0.0
tensorflow>=2.10.0
numpy>=1.23.0
Pillow>=9.0.0
opencv-python>=4.8.0
scikit-learn>=1.2.0
joblib>=1.2.0
reportlab>=4.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0
